+++ xxtea.c
@@ -32,7 +32,7 @@
 #include <stdint.h>
 #include <string.h>
 
-#define VERSION "6.0.0"
+#define VERSION "6.1.0"
 
 enum {
     XXTEA_PADDING_NONE = 0,